+++ pkg/run/bootstrap/bootstrap_wizard_ui.go
@@ -1,4 +1,4 @@
-package run
+package bootstrap
 
 import (
 	"fmt"
