+++ pkg/run/install/apply_manifests.go
@@ -1,4 +1,4 @@
-package run
+package install
 
 import (
 	"bytes"
