+++ pkg/run/install/apply_manifests_test.go
@@ -1,4 +1,4 @@
-package run
+package install
 
 import (
 	"context"
