+++ pkg/run/install/install_dashboard.go
@@ -1,4 +1,4 @@
-package run
+package install
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 	helmv2 "github.com/fluxcd/helm-controller/api/v2beta1"
 	sourcev1 "github.com/fluxcd/source-controller/api/v1beta2"
 	"github.com/weaveworks/weave-gitops/pkg/logger"
-	"github.com/weaveworks/weave-gitops/pkg/server"
+	"github.com/weaveworks/weave-gitops/pkg/run"
 	"github.com/weaveworks/weave-gitops/pkg/utils"
 	"golang.org/x/crypto/bcrypt"
 	corev1 "k8s.io/api/core/v1"
@@ -18,7 +18,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/kustomize/api/resource"
 	kyaml "sigs.k8s.io/kustomize/kyaml/yaml"
@@ -113,47 +112,6 @@ func getDashboardHelmChart(log logger.Logger, ctx context.Context, kubeClient cl
 	return &helmChart
 }
 
-// EnablePortForwardingForDashboard enables port forwarding for the GitOps Dashboard.
-func EnablePortForwardingForDashboard(log logger.Logger, kubeClient client.Client, config *rest.Config, namespace string, podName string, dashboardPort string) (func(), error) {
-	specMap := &PortForwardSpec{
-		Namespace:     namespace,
-		Name:          podName,
-		Kind:          "deployment",
-		HostPort:      dashboardPort,
-		ContainerPort: server.DefaultPort,
-	}
-	// get pod from specMap
-	namespacedName := types.NamespacedName{Namespace: specMap.Namespace, Name: specMap.Name}
-
-	pod, err := GetPodFromResourceDescription(namespacedName, specMap.Kind, kubeClient)
-	if err != nil {
-		log.Failuref("Error getting pod from specMap: %v", err)
-	}
-
-	if pod != nil {
-		waitFwd := make(chan struct{}, 1)
-		readyChannel := make(chan struct{})
-		cancelPortFwd := func() {
-			close(waitFwd)
-		}
-
-		log.Actionf("Port forwarding to pod %s/%s ...", pod.Namespace, pod.Name)
-
-		go func() {
-			if err := ForwardPort(pod, config, specMap, waitFwd, readyChannel); err != nil {
-				log.Failuref("Error forwarding port: %v", err)
-			}
-		}()
-		<-readyChannel
-
-		log.Successf("Port forwarding for dashboard is ready.")
-
-		return cancelPortFwd, nil
-	}
-
-	return nil, ErrDashboardPodNotFound
-}
-
 // ReconcileDashboard reconciles the dashboard.
 func ReconcileDashboard(kubeClient client.Client, name string, namespace string, podName string, timeout time.Duration) error {
 	const interval = 3 * time.Second / 2
@@ -175,7 +133,7 @@ func ReconcileDashboard(kubeClient client.Client, name string, namespace string,
 
 	if err := wait.Poll(interval, timeout, func() (bool, error) {
 		var err error
-		sourceRequestedAt, err = requestReconciliation(context.Background(), kubeClient,
+		sourceRequestedAt, err = run.RequestReconciliation(context.Background(), kubeClient,
 			namespacedName, gvk)
 
 		return err == nil, nil
@@ -202,7 +160,7 @@ func ReconcileDashboard(kubeClient client.Client, name string, namespace string,
 	if err := wait.Poll(interval, timeout, func() (bool, error) {
 		namespacedName := types.NamespacedName{Namespace: namespace, Name: podName}
 
-		dashboard, _ := GetPodFromResourceDescription(namespacedName, "deployment", kubeClient)
+		dashboard, _ := run.GetPodFromResourceDescription(namespacedName, "deployment", kubeClient)
 		if dashboard == nil {
 			return false, nil
 		}
