+++ pkg/run/install/install_dashboard_test.go
@@ -1,4 +1,4 @@
-package run
+package install
 
 import (
 	"context"
