+++ pkg/run/install/install_flux.go
@@ -1,4 +1,4 @@
-package run
+package install
 
 import (
 	"context"
