+++ pkg/run/install/install_suite_test.go
@@ -1,0 +1,34 @@
+package install
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"github.com/weaveworks/weave-gitops/pkg/testutils"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+var (
+	k8sClient client.Client
+	k8sEnv    *testutils.K8sTestEnv
+)
+
+func TestRun(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Run Suite")
+}
+
+var cleanupK8s func()
+
+var _ = BeforeSuite(func() {
+	var err error
+	k8sEnv, err = testutils.StartK8sTestEnvironment([]string{
+		"../../manifests/crds",
+		"../../tools/testcrds",
+	})
+	Expect(err).NotTo(HaveOccurred())
+
+	cleanupK8s = k8sEnv.Stop
+	k8sClient = k8sEnv.Client
+})
