+++ pkg/run/watch/forward_port.go
@@ -1,22 +1,16 @@
-package run
+package watch
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
 	"os"
 	"strings"
 
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/portforward"
 	"k8s.io/client-go/transport/spdy"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 type PortForwardSpec struct {
@@ -115,74 +109,4 @@ func ForwardPort(pod *corev1.Pod, cfg *rest.Config, specMap *PortForwardSpec, wa
 	}
 
 	return fw.ForwardPorts()
-}
-
-func GetPodFromResourceDescription(namespacedName types.NamespacedName, kind string, kubeClient client.Client) (*corev1.Pod, error) {
-	switch kind {
-	case "pod":
-		pod := &corev1.Pod{}
-		if err := kubeClient.Get(context.Background(), namespacedName, pod); err != nil {
-			return nil, err
-		}
-
-		return pod, nil
-	case "service":
-		svc := &corev1.Service{}
-		if err := kubeClient.Get(context.Background(), namespacedName, svc); err != nil {
-			return nil, fmt.Errorf("error getting service: %s, namespaced Name: %v", err, namespacedName)
-		}
-
-		// list pods of the service "svc" by selector in a specific namespace using the controller-runtime client
-		podList := &corev1.PodList{}
-		if err := kubeClient.List(context.Background(), podList,
-			client.MatchingLabelsSelector{
-				Selector: labels.Set(svc.Spec.Selector).AsSelector(),
-			},
-			client.InNamespace(svc.Namespace),
-		); err != nil {
-			return nil, err
-		}
-
-		if len(podList.Items) == 0 {
-			return nil, ErrNoPodsForService
-		}
-
-		for _, pod := range podList.Items {
-			if pod.Status.Phase == corev1.PodRunning {
-				return &pod, nil
-			}
-		}
-
-		return nil, ErrNoRunningPodsForService
-	case "deployment":
-		deployment := &appsv1.Deployment{}
-		if err := kubeClient.Get(context.Background(), namespacedName, deployment); err != nil {
-			return nil, fmt.Errorf("error getting deployment: %s, namespaced Name: %v", err, namespacedName)
-		}
-
-		// list pods of the deployment "deployment" by selector in a specific namespace using the controller-runtime client
-		podList := &corev1.PodList{}
-		if err := kubeClient.List(context.Background(), podList,
-			client.MatchingLabelsSelector{
-				Selector: labels.Set(deployment.Spec.Selector.MatchLabels).AsSelector(),
-			},
-			client.InNamespace(deployment.Namespace),
-		); err != nil {
-			return nil, err
-		}
-
-		if len(podList.Items) == 0 {
-			return nil, ErrNoPodsForDeployment
-		}
-
-		for _, pod := range podList.Items {
-			if pod.Status.Phase == corev1.PodRunning {
-				return &pod, nil
-			}
-		}
-
-		return nil, ErrNoRunningPodsForDeployment
-	}
-
-	return nil, errors.New("unsupported spec kind")
 }
